refactor: 메인페이지 구동 및 로그인 페이지 연결
