+++ src/main/webapp/index.jsp
@@ -11,7 +11,7 @@
 </h1>
 
 <%
-    response.sendRedirect("login.do");
+    response.sendRedirect("main.do");
 %>
 
 </body>
